requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
crontab==1.0.1
Flask==2.3.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.5
requests==2.30.0
six==1.16.0
sseclient==0.0.27
urllib3==2.0.3
Werkzeug==2.3.7
zipp==3.16.2
